added float input support

+++ pypy/translator/goal/targetgbfullprofiling.py
@@ -8,7 +8,7 @@ ROM_PATH = str(py.magic.autopath().dirpath().dirpath().dirpath())+"/lang/gameboy
 def entry_point(argv=None):
     if argv is not None and len(argv) > 1:
         filename = argv[1]
-        execution_seconds = int(argv[2])
+        execution_seconds = float(argv[2])
     else:
         pos = str(9)
         filename = ROM_PATH+"/rom"+pos+"/rom"+pos+".gb"
